pyproject.toml
[project]
name = "ai-agent-from-scratch"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai>=1.50.0",
    "anthropic>=0.40.0",
    "litellm>=1.50.0",
    "pydantic>=2.0",
    "python-dotenv>=1.0.0",
    "tavily-python>=0.5.0",
    "mcp>=1.0.0",
    "fastmcp>=0.1.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
    "tiktoken>=0.7.0",
    "chromadb>=0.5.0",
    "e2b-code-interpreter>=1.0.0",
    "pandas>=2.0",
    "openpyxl>=3.1.0",
    "pymupdf>=1.24.0",
    "httpx>=0.27.0",
    "datasets>=2.0",
    "tqdm>=4.0",
    "marimo",
    "python-docx>=1.0.0",
]
